+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Sort files by width and height instead of file size
+
 ## [0.12.1] - 2023-03-22
 
 ### Fixed